Add files via upload
"This repository contains the Hangman game I implemented as part of the MIT Introduction to Computer Science and Programming Using Python course. In this project, I've applied the concepts of functions, loops, and conditionals in Python to create an interactive command-line game. The user has a set number of attempts to guess a secret word letter by letter. This project demonstrates my understanding of Python basics and my ability to create a simple yet fun application. I welcome any feedback or suggestions!"